requirements.txt
# Core data science
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.2.0


# Model export
onnx>=1.14.0
skl2onnx>=1.15.0
